end of back

+++ src/controllers/videoController.js
@@ -1,4 +1,6 @@
-import Video from "../models/Video";
+import User from "../models/User"
+import Video from "../models/Video"
+
 
 export const home = (req,res) => {
     Video.find()
@@ -17,29 +19,42 @@ export const home = (req,res) => {
 
 export const watch = async (req, res) => {
     const {id} = req.params;
-    const video = await Video.findById(id); 
+    const video = await Video.findById(id).populate("owner");
     if(!video){
         return res.status(404).render("404", {pageTitle:" Video not found."})   
     }
-    return res.render("watch", { pageTitle: video.title, video:video });
+
+    return res.render("watch", { pageTitle: video.title, video:video,  });
 };
 
 export const getEdit = async (req, res) => {
     const {id} = req.params;
+    const{
+        user: { _id },
+    } = req.session;
     const video = await Video.findById(id); 
     if(!video){
         return res.status(404).render("404", {pageTitle:" Video not found."})   
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/")
     }
     return res.render("edit", { pageTitle: `Edit : ${video.title}`, video:video });
 };
 
 export const postEdit = async (req, res) => {
+    const{
+        user: { _id },
+    } = req.session;
     const {id} = req.params;
     const {title, description, hashtags} = req.body;
     const video = await Video.exists({ _id: id}); 
     //.exists requires filter
     if(!video){
         return res.render("404", {pageTitle:" Video not found."});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/")
     }
     await Video.findByIdAndUpdate(id, {
         title: title,
@@ -55,14 +70,23 @@ export const getUpload = (req,res) => {
 
 export const postUpload = async (req, res) => {
     //here we will add a video to the videos array.
+    const {
+        user: {_id},
+    } = req.session;
+    const {path} = req.file;
     const {title, description, hashtags} = req.body;
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title: title,
             description: description,
+            fileUrl: path,
+            owner:_id,
             createdAt: Date.now(),
             hashtags:Video.formatHashtags(hashtags),
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/");
     } catch (error) {
         return res.status(400).render("upload", {
@@ -74,8 +98,17 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req,res) => {
     const { id } = req.params;
+    const {
+        user: {_id},
+    } = req.session;
+    const video = await Video.findById(id);
+    if(!video) {
+        return res.status(404).render("404", {pageTitle:" Video not found."})  
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/")
+    }
     await Video.findByIdAndDelete(id);
-    console.log(id);
     return res.redirect("/");
 };
 
@@ -88,6 +121,6 @@ export const search = async (req,res) => {
                 $regex: new RegExp(keyword, "i")
             },
         });
-    }
+    };
     return res.render("search", {pageTitle: "Search", videos});
 };
